[WIP] target: avr32_ap7k: Add timeouts to infinite loops
The only way an operation could take more than a few dozen
milliseconds for the target to process is if we're sending USB packets
from a VM through a spice channel. The more common failure modes are
the chip not responding, or a problem with the JTAG connectoion.

In either case, the 1 second timeout is more than generous to cover
all cases, but also doesn't block execution in case something is
really wrong.