avr32: avr32_jtag_write_reg: Do not wait for EMUD (dirty bit)
On AVR32, writing to a CPU register via JTAG is achieved in two steps.
  1. The DCEMU JTAG register is programmed with the value to load.
   - The MFDR (move from debug register) is then loaded to DCINST register.
  2. One CPU instruction is executed to load the value in a CPU register.

Once DCEMU is written, the EMUD bit in DCSR is set by hardware. The bit is
then cleared automatically when the CPU reads DCEMU. Thus the while loop is
wrong. It should wait for EMUD to clear, not to get set.

However, We know the CPU has read DCEMU because avr32_jtag_exec() checks
the DS register for the instruction complete bit (INC). Thus, waiting for
EMUD to be cleared is redundant.

Atmel Studio also does not check DCSR in this operation.

Change-Id: I02bc393f49fa2cb42cd4818096923b969524b8a9